New PID configuration API to improve ease-of-use
This commit introduces a new library interface to configure PID properties with
greater ease-of-use and developer experience. This change allows this Rust
based implementation to become more approachable for both new and existing
developers looking to integrate such software into their ecosystems, providing
a wider circulation for this library.

API is incompatible with v3 but has functional parity.